@@ -1,46 +1,59 @@
 package utils
 
 import (
+	"fmt"
 	"io"
-	"io/ioutil"
 	"net/http"
 	"net/http/cookiejar"
+	"strconv"
 	"time"
 
 	"github.com/sirupsen/logrus"
 )
 
+// Log is the logger for the http client
 var Log = logrus.New()
 
-/*
-Notes to myself: Figured out pretty much how I'm going to develop inhertiance fo the
-lower level HTTP requests, I have the functions copied to this struct right now, just
-need to edit Put and Delete to account for the syntax of the error struct in here.
-The way I'm going to develop inheritance is most likely going to be by composition, by having
-this client referenced in each client, then having their methods called,sort of like
-how *http.Client is created in here, though I think we don't even need that in the other
-ones which seems interesting.
-*/
-
-//change name of Client since it might mess with the inheritance for the other clients
+// change name of Client since it might mess with the inheritance for the other clients
 type Client struct {
 	client  *http.Client
-	url     string
+	urlBase string
 	timeout int
 }
 
-func (c *Client) IsConnected() bool {
-	return c.client != nil
+// IsConnected checks if the client has successfully connected to the specified url via a TRACE request
+func (c *Client) IsConnected() (bool, string) {
+	traceRequest, err := http.NewRequest(http.MethodTrace, c.urlBase+"/", nil)
+	if err != nil {
+		return false, err.Error()
+	}
+
+	res, err := c.client.Do(traceRequest)
+	if err != nil {
+		return false, err.Error()
+	}
+
+	if res.StatusCode != 200 {
+		return false, "ERROR: status code " + strconv.Itoa(res.StatusCode)
+	}
+
+	return true, ""
 }
 
-func NewClient(url string, timeout int) *Client {
+// NewClient creates an HTTP client to interact with an HTTP server
+// at a specified URL.
+func NewClient(urlBase string, timeout int) *Client {
 	client := &Client{
 
-		url:     "http://" + url,
+		urlBase: "http://" + urlBase,
 		timeout: timeout,
 	}
 
-	cookieJar, _ := cookiejar.New(nil)
+	cookieJar, err := cookiejar.New(nil)
+	if err != nil {
+		Log.Debugf("Could not create client cookie jar. Reason: %s", err)
+	}
+
 	client.client = &http.Client{
 		Jar:     cookieJar,
 		Timeout: time.Duration(timeout) * time.Second,
@@ -49,14 +62,11 @@ func NewClient(url string, timeout int) *Client {
 	return client
 }
 
-//TODO - find a wway to rename the outputed strings to not be specific to a server but generalized, or have a way for
-//the clients to properly account for them
-
-//Post makes a POST request to the server
+// Post makes a POST request to the server
 func (c *Client) Post(uri string, msg io.Reader) ([]byte, HTTPError) {
 	httpErr := NewHTTPError()
 
-	resp, err := c.client.Post(c.url+uri, "application/json", msg)
+	resp, err := c.client.Post(c.urlBase+uri, "application/json", msg)
 
 	// If err is not nil, then the server is not online and we need to back out
 	// so that nothing crashes
@@ -68,23 +78,26 @@ func (c *Client) Post(uri string, msg io.Reader) ([]byte, HTTPError) {
 
 	// The server is online, but we need to check the status code directly to
 	// see if there was a 4xx error
+	errMsg, err := io.ReadAll(resp.Body)
 	if resp.StatusCode != http.StatusOK {
-		errMsg, _ := ioutil.ReadAll(resp.Body)
+		if err != nil {
+			errMsg = []byte("Unknown error message")
+		}
+
 		httpErr.SetError("POST", errMsg, resp.StatusCode)
 	}
 
 	defer resp.Body.Close()
-	body, _ := ioutil.ReadAll(resp.Body)
 	Log.Debugf("Making Request : POST - %s - %d", uri, resp.StatusCode)
 
-	return body, *httpErr
+	return errMsg, *httpErr
 }
 
-//Get makes a GET request to the server
+// Get makes a GET request to the server
 func (c *Client) Get(uri string) ([]byte, HTTPError) {
 	httpErr := NewHTTPError()
 
-	resp, err := c.client.Get(c.url + uri)
+	resp, err := c.client.Get(c.urlBase + uri)
 
 	if err != nil {
 		Log.Warn(err)
@@ -93,11 +106,14 @@ func (c *Client) Get(uri string) ([]byte, HTTPError) {
 	}
 
 	if resp.StatusCode != http.StatusOK {
-		errMsg, _ := ioutil.ReadAll(resp.Body)
+		errMsg, err := io.ReadAll(resp.Body)
+		if err != nil {
+			errMsg = []byte("Unknown error message")
+		}
 		httpErr.SetError("GET", errMsg, resp.StatusCode)
 	}
 	defer resp.Body.Close()
-	body, resErr := ioutil.ReadAll(resp.Body)
+	body, resErr := io.ReadAll(resp.Body)
 	if resErr != nil {
 		Log.Debug(resErr)
 	}
@@ -112,7 +128,12 @@ func (c *Client) Put(uri string, msg io.Reader) ([]byte, HTTPError) {
 	httpErr := NewHTTPError()
 
 	// set the HTTP method, url, and request body
-	req, _ := http.NewRequest(http.MethodPut, c.url+uri, msg)
+	req, err := http.NewRequest(http.MethodPut, c.urlBase+uri, msg)
+	if err != nil {
+		Log.Debug(err)
+		httpErr.SetError("PUT", []byte(fmt.Sprintf("Could not create request. Reason: %s", err)), http.StatusInternalServerError)
+		return nil, *httpErr
+	}
 
 	// set the request header Content-Type for json
 	req.Header.Set("Content-Type", "application/json")
@@ -128,15 +149,17 @@ func (c *Client) Put(uri string, msg io.Reader) ([]byte, HTTPError) {
 
 	// The server is online, but we need to check the status code directly to
 	// see if there was a 4xx error
+	errMsg, err := io.ReadAll(resp.Body)
 	if resp.StatusCode != http.StatusOK {
-		errMsg, _ := ioutil.ReadAll(resp.Body)
+		if err != nil {
+			errMsg = []byte("Unknown error message")
+		}
+
 		httpErr.SetError("PUT", errMsg, resp.StatusCode)
 	}
 
-	body, _ := ioutil.ReadAll(resp.Body)
-
 	Log.Debugf("Making Request to: PUT - %s - %d", uri, resp.StatusCode)
-	return body, *httpErr
+	return errMsg, *httpErr
 }
 
 // Delete makes a DELETE request to the server
@@ -144,7 +167,13 @@ func (c *Client) Delete(uri string) ([]byte, HTTPError) {
 	httpErr := NewHTTPError()
 
 	// set the HTTP method, url, and request body
-	req, err := http.NewRequest(http.MethodDelete, c.url+uri, nil)
+	req, err := http.NewRequest(http.MethodDelete, c.urlBase+uri, nil)
+	if err != nil {
+		Log.Debug(err)
+		httpErr.SetError("DELETE", []byte(fmt.Sprintf("Could not create request. Reason: %s", err)), http.StatusInternalServerError)
+		return nil, *httpErr
+	}
+
 	resp, err := c.client.Do(req)
 
 	// If err is not nil, then the server is not online and we need to back out
@@ -157,12 +186,21 @@ func (c *Client) Delete(uri string) ([]byte, HTTPError) {
 
 	// The server is online, but we need to check the status code directly to
 	// see if there was a 4xx error
+	errMsg, err := io.ReadAll(resp.Body)
 	if resp.StatusCode != 200 {
-		errMsg, _ := ioutil.ReadAll(resp.Body)
+		if err != nil {
+			errMsg = []byte("Unknown error message")
+		}
+
 		httpErr.SetError("DELETE", errMsg, resp.StatusCode)
 	}
 
-	body, _ := ioutil.ReadAll(resp.Body)
+	body, err := io.ReadAll(resp.Body)
+	if err != nil {
+		Log.Debug(err)
+		httpErr.SetError("DELETE", []byte("Could not read response body"), http.StatusInternalServerError)
+		return nil, *httpErr
+	}
 
 	Log.Debugf("Making Request to: DELETE - %s - %d", uri, resp.StatusCode)
 	return body, *httpErr
