@@ -1,1061 +1,900 @@
 package server
 
 import (
-	"bytes"
-	"context"
-	"encoding/base64"
-	"encoding/json"
 	"fmt"
-	"io/ioutil"
 	"net/http"
 	"strconv"
 	"strings"
 	"time"
 
-	influxdb2 "github.com/influxdata/influxdb-client-go/v2"
-	"github.com/rs/cors"
+	"github.com/gin-contrib/static"
+	"github.com/gin-gonic/gin"
+
 	"github.com/sirupsen/logrus"
-
-	cv "github.com/tritonuas/hub/internal/computer_vision"
-	ic "github.com/tritonuas/hub/internal/interop"
-
-	// mav "github.com/tritonuas/hub/internal/mavlink"
-	pp "github.com/tritonuas/hub/internal/path_plan"
-	ut "github.com/tritonuas/hub/internal/utils"
+	"github.com/tritonuas/gcs/internal/cvs"
+	"github.com/tritonuas/gcs/internal/influxdb"
+	"github.com/tritonuas/gcs/internal/manager"
+	mav "github.com/tritonuas/gcs/internal/mavlink"
+	"github.com/tritonuas/gcs/internal/obc"
+	"github.com/tritonuas/gcs/internal/obc/airdrop"
+	"github.com/tritonuas/gcs/internal/obc/pp"
 )
 
+// Log is the logger for the server
 var Log = logrus.New()
 
-// https://app.clickup.com/t/28rwhv5
-
-// Server provides the implementation for the hub server that communicates
-// with other parts of the plane's system and houston
+/*
+Stores the server state and data that the server deals with.
+*/
 type Server struct {
-	port               string
-	client             *ic.Client
-	pathPlanningClient *pp.Client
-	cvData             *cv.ComputerVisionData
-
-	telemetry []byte // Holds the most recent telemetry data sent to the interop server
-
-	path *pp.Path // Holds the path of the plane, see the definition of the struct for more details
-
-	homePosition *ic.Position // Home position of the plane, which must be set by us
-
-	missionID MissionID
-
-	gcsWaypoint map[string]float64
-
-	//mission TODO Actually hold the mission object for pyplanner to request
+	influxDBClient      *influxdb.Client
+	mavlinkClient       *mav.Client
+	obcClient           *obc.Client
+	UnclassifiedTargets []cvs.UnclassifiedODLC `json:"unclassified_targets"`
+	Bottles             *airdrop.Bottles
+	MissionTime         int64
+	FlightBounds        []pp.Coordinate
+	AirDropBounds       []pp.Coordinate
+	ClassifiedTargets   []cvs.ClassifiedODLC
+	Manager             *manager.Manager
 }
 
-// Run starts the hub http server and establishes all of the uri's that it
-// will receive
-func (s *Server) Run(
-	port string,
-	interopChannel chan *ic.Client,
-	interopMissionID int,
-	rtppChannel chan *pp.Client,
-	telemetryChannel chan *ic.Telemetry,
-	influxdbURI string,
-	influxToken string,
-	influxBucket string,
-	influxOrg string,
-	sendWaypointToPlaneChannel chan *pp.Path) {
+/*
+We aren't hosting this online, so it's okay to allow requests from all origins to make Houston2's life easier
+*/
+func CORSMiddleware() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		c.Writer.Header().Set("Access-Control-Allow-Origin", "*")
+		c.Writer.Header().Set("Access-Control-Allow-Credentials", "true")
+		c.Writer.Header().Set("Access-Control-Allow-Headers", "Content-Type, Content-Length, Accept-Encoding, X-CSRF-Token, Authorization, accept, origin, Cache-Control, X-Requested-With")
+		c.Writer.Header().Set("Access-Control-Allow-Methods", "POST, OPTIONS, GET, PUT")
 
-	// receives "true" when the mission is started
-	missionStartChannel := make(chan bool)
+		if c.Request.Method == "OPTIONS" {
+			c.AbortWithStatus(204)
+			return
+		}
 
-	s.missionID = MissionID{ID: interopMissionID}
+		c.Next()
+	}
+}
 
-	s.pathPlanningClient = pp.NewClient("127.0.0.1:5000", 5)
-
-	s.cvData = cv.InitializeData()
-
-	s.port = fmt.Sprintf(":%s", port)
-	s.client = nil
-	go s.ConnectToInterop(interopChannel)
-	go s.ConnectToRTPP(rtppChannel)
-	mux := http.NewServeMux()
-	mux.Handle("/hub/gcs", &gcsPositionHandler{server: s, uri: influxdbURI, token: influxToken, bucket: influxBucket, org: influxOrg})
-
-	mux.Handle("/hub/interop/teams", &interopTeamHandler{server: s})       // get info about teams from interop
-	mux.Handle("/hub/interop/missions", &interopMissionHandler{server: s}) // get mission from interop using server's mission ID
-	mux.Handle("/hub/interop/telemetry", &interopTelemHandler{server: s})  // get other teams telem info from interop
-
-	mux.Handle("/hub/mission/id", &missionHandler{server: s}) // GET: get current id we're using
-	mux.Handle("/hub/mission/initialize", &missionHandlerInitialize{server: s})
-	mux.Handle("/hub/mission/start", &missionHandlerStart{server: s, waypointChan: sendWaypointToPlaneChannel, missionStartChan: missionStartChannel})
-
-	mux.Handle("/hub/path", &pathCacherHandler{server: s})
-
-	mux.Handle("/hub/plane/home", &planeHomeHandler{server: s})
-	mux.Handle("/hub/plane/telemetry", &planeTelemetryHandler{server: s, uri: influxdbURI, token: influxToken, bucket: influxBucket, org: influxOrg})
-
-	mux.Handle("/hub/interop/odlc/", &interopOdlcHandler{server: s})
-	mux.Handle("/hub/interop/odlcs", &interopOdlcsHandler{server: s})
-	mux.Handle("/hub/interop/odlc/image/", &interopOdlcImageHandler{server: s})
-
-	mux.Handle("/hub/cv/cropped/", &CVCroppedHandler{server: s, uri: influxdbURI, token: influxToken, bucket: influxBucket, org: influxOrg}) // eventually these influx variables will be in an influxclient struct
-	mux.Handle("/hub/cv/result/", &CVResultHandler{server: s})
-
-	c := cors.New(cors.Options{
-		AllowedMethods: []string{"GET", "POST", "PUT", "DELETE"},
+func (server *Server) initFrontend(router *gin.Engine) {
+	router.Use(static.Serve("/", static.LocalFile("static", false)))
+	router.NoRoute(func(c *gin.Context) {
+		c.Redirect(http.StatusTemporaryRedirect, "/404.html")
 	})
-
-	go s.CacheAndUploadTelem(influxdbURI, influxToken, influxBucket, influxOrg, missionStartChannel)
-	handler := c.Handler(mux)
-	http.ListenAndServe(s.port, handler)
 }
 
-func (s *Server) ConnectToInterop(channel chan *ic.Client) {
-	s.client = <-channel
-	Log.Info("Server client object initialized: Interop fully connected.")
-}
+func (server *Server) initBackend(router *gin.Engine) {
+	api := router.Group("/api")
+	{
+		/*
+			TODO:
 
-func (s *Server) ConnectToRTPP(channel chan *pp.Client) {
-	s.pathPlanningClient = <-channel
-	Log.Info("Server client object initialized: RTPP fully connected.")
-}
+			---------------------------
 
-// CacheAndUploadTelem queries the database for telemetry and posts it to the interop server
-func (s *Server) CacheAndUploadTelem(uri string, token string, bucket string, org string, missionStartChannel chan bool) {
-	client := influxdb2.NewClient(uri, token)
-	queryAPI := client.QueryAPI(org)
+			1. in the mission manager, detect a change in state to airdrop and forward the classified targets to obc
+		*/
 
-	latQueryString := fmt.Sprintf(`from(bucket:"%s") |> range(start: -1m) |> tail(n: 1, offset: 0) |> filter(fn: (r) => r.ID == "33") |> filter(fn: (r) => r._field == "lat")`, bucket)
-	lonQueryString := fmt.Sprintf(`from(bucket:"%s") |> range(start: -1m) |> tail(n: 1, offset: 0) |> filter(fn: (r) => r.ID == "33") |> filter(fn: (r) => r._field == "lon")`, bucket)
-	altQueryString := fmt.Sprintf(`from(bucket:"%s") |> range(start: -1m) |> tail(n: 1, offset: 0) |> filter(fn: (r) => r.ID == "33") |> filter(fn: (r) => r._field == "alt")`, bucket)
-	hdgQueryString := fmt.Sprintf(`from(bucket:"%s") |> range(start: -1m) |> tail(n: 1, offset: 0) |> filter(fn: (r) => r.ID == "33") |> filter(fn: (r) => r._field == "hdg")`, bucket)
+		api.GET("/connections", server.testConnections())
 
-	queryStrings := make([]string, 4)
-	queryStrings[0] = latQueryString
-	queryStrings[1] = lonQueryString
-	queryStrings[2] = altQueryString
-	queryStrings[3] = hdgQueryString
-
-	fields := make([]string, 4)
-	fields[0] = "latitude"
-	fields[1] = "longitude"
-	fields[2] = "altitude"
-	fields[3] = "heading"
-
-	Log.Info("Waiting for mission to start to begin uploaded telemetry to the interop server.")
-	_ = <-missionStartChannel
-	Log.Info("Mission has been started so we will begin to upload telemetry to the interop server.")
-
-loop:
-	for true {
-		// Note: this code is basically copied from /hub/plane/telemetry endpoint.
-		// In the future we should abstract out this logic into a separate module that just deals with interfacing with the influx db
-		// so that this code and the code in /hub/plane/telemtry just uses this module instead of having to do the dirty work itself
-		var results []string
-		for _, queryString := range queryStrings {
-			result, err := queryAPI.Query(context.Background(), queryString)
-			if err != nil {
-				Log.Errorf("Error querying Influx for telemetry data: %s ", err.Error())
-				continue loop
-			} else {
-				if result.Next() {
-					val := fmt.Sprint(result.Record().Value())
-					results = append(results, val)
-				} else {
-					Log.Error("Error querying Influx for telemetry data: no value found")
-					continue loop
-				}
-			}
-		}
-		jsonMap := make(map[string]float64)
-		for i, field := range fields {
-			val, _ := strconv.ParseFloat(results[i], 64)
-			if field == "latitude" || field == "longitude" {
-				val /= 1e7
-			}
-			if field == "altitude" {
-				val /= 1000
-			}
-			if field == "heading" {
-				val /= 100
-			}
-			jsonMap[field] = val
-		}
-
-		jsonStr, _ := json.Marshal(jsonMap)
-
-		// TODO: consider putting a rate limit on this so we don't spam the interop server?
-		if s.client != nil && s.client.IsConnected() {
-			s.client.PostTelemetry(jsonStr)
-		}
-
-		time.Sleep(500 * time.Millisecond)
-	}
-}
-
-func logRequestInfo(r *http.Request) {
-	Log.Infof("Request to Hub from %s: %s %s", r.RemoteAddr, r.Method, r.URL)
-}
-
-type gcsPositionHandler struct {
-	server *Server
-	uri    string
-	token  string
-	org    string
-	bucket string
-}
-
-func (m *gcsPositionHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-	logRequestInfo(r)
-	switch r.Method {
-	case "POST":
-		client := influxdb2.NewClient(m.uri, m.token)
-		queryAPI := client.QueryAPI(m.org)
-
-		latQueryString := fmt.Sprintf(`from(bucket:"%s") |> range(start: -1m) |> tail(n: 1, offset: 0) |> filter(fn: (r) => r.ID == "33") |> filter(fn: (r) => r._field == "lat")`, m.bucket)
-		lonQueryString := fmt.Sprintf(`from(bucket:"%s") |> range(start: -1m) |> tail(n: 1, offset: 0) |> filter(fn: (r) => r.ID == "33") |> filter(fn: (r) => r._field == "lon")`, m.bucket)
-
-		queryStrings := make([]string, 2)
-		queryStrings[0] = latQueryString
-		queryStrings[1] = lonQueryString
-
-		fields := make([]string, 2)
-		fields[0] = "latitude"
-		fields[1] = "longitude"
-
-		// Log.Info("Waiting for mission to start to begin uploaded telemetry to the interop server.")
-		// _ = <-m.server.missionStartChannel
-		// Log.Info("Mission has been started so we will begin to upload telemetry to the interop server.")
-
-		// Note: this code is basically copied from /hub/plane/telemetry endpoint.
-		// In the future we should abstract out this logic into a separate module that just deals with interfacing with the influx db
-		// so that this code and the code in /hub/plane/telemtry just uses this module instead of having to do the dirty work itself
-		var results []string
-		for _, queryString := range queryStrings {
-			result, err := queryAPI.Query(context.Background(), queryString)
-			if err != nil {
-				w.WriteHeader(http.StatusInternalServerError)
-				w.Write([]byte(fmt.Sprintf("Error Querying InfluxDB: %s", err)))
-				return
-			} else {
-				if result.Next() {
-					val := fmt.Sprint(result.Record().Value())
-					results = append(results, val)
-				} else {
-					w.WriteHeader(http.StatusBadRequest)
-					w.Write([]byte(fmt.Sprintf("Requested telemetry with query %s not found in InfluxDB. Check the id and field in the Mavlink documentation at http://mavlink.io/en/messages/common.html", queryString)))
-					return
-				}
-			}
-		}
-		jsonMap := make(map[string]float64)
-		for i, field := range fields {
-			val, _ := strconv.ParseFloat(results[i], 64)
-			if field == "latitude" || field == "longitude" {
-				val /= 1e7
-			}
-			jsonMap[field] = val
-		}
-
-		m.server.gcsWaypoint = jsonMap
-		client.Close()
-	default:
-		w.WriteHeader(http.StatusNotImplemented)
-		w.Write([]byte("Not Implemented"))
-	}
-}
-
-type missionHandler struct {
-	server *Server
-}
-
-// This object captures changes to the mission ID stored in Hub
-// To change the mission ID that hub is using:
-// POST /interop/missions
-// {
-//  	"id": [MISSION_ID]
-// }
-
-// MissionID is an object used to capture a mission ID parameter
-type MissionID struct {
-	ID int `json:"id,omitempty"`
-}
-
-func (m *missionHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-	logRequestInfo(r)
-
-	switch r.Method {
-	case "GET":
-		idData, err := json.Marshal(m.server.missionID)
-		if err != nil {
-			w.WriteHeader(http.StatusBadRequest)
-			//w.Write(err.Message)
-		} else {
-			w.Write(idData)
-			Log.Infof("Successfully retrieved mission ID information: id = %d", m.server.missionID)
-		}
-
-	default:
-		w.WriteHeader(http.StatusNotImplemented)
-		w.Write([]byte("Not implemented"))
-	}
-}
-
-type missionHandlerInitialize struct {
-	server *Server
-}
-
-func (m missionHandlerInitialize) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-	//cut path_plan/initialize code
-	logRequestInfo(r)
-
-	if m.server.pathPlanningClient == nil {
-		w.WriteHeader(http.StatusInternalServerError)
-		w.Write([]byte("Path Planning connection not established"))
-		Log.Errorf("Unable to get data from Path Planning; connection not established")
-		return
-	}
-
-	switch r.Method {
-	case "POST":
-		//use post function from pp client
-		missionID, _ := ioutil.ReadAll(r.Body) //reads mission ID from post request body
-		//err := ic.NewInteropError()
-
-		var id MissionID
-		jsonErr := json.Unmarshal(missionID, &id)
-
-		if jsonErr != nil {
-			w.WriteHeader(http.StatusBadRequest)
-			w.Write([]byte(jsonErr.Error()))
-			w.Write(missionID)
-			break
-		}
-
-		m.server.missionID = id
-		fmt.Println(id.ID)
-
-		missionData, err := m.server.client.GetMission(id.ID) //gets mission data from interop server
-
-		if err.Post {
-			w.WriteHeader(err.Status)
-			w.Write(err.Message)
-			Log.Errorf("Unable to send data to path planning: %s", err.Message)
-			break
-		} else {
-			Log.Infof("Successfully requested mission information from interop")
-		}
-
-		ppErr := m.server.pathPlanningClient.PostMission(missionData) //calls pp client post function
-
-		if ppErr.Post {
-			w.WriteHeader(ppErr.Status)
-			w.Write(ppErr.Message)
-			Log.Errorf("Unable to post mission data to path planning: %s", ppErr.Message)
-			break
-		}
-
-		//copied path_plan/plath code for 3d sub bullet
-
-		//Make the GET request to the PathPlan Server
-		//path, err := json.Marhsall(p.path.Waypoint)
-		path, pathBinary, err := m.server.pathPlanningClient.GetPath()
-		if err.Get {
-			w.WriteHeader(err.Status)
-			w.Write(err.Message)
-		} else {
-			m.server.path = &path //caching the path
-			w.Write(pathBinary)
-		}
-	default:
-		w.WriteHeader(http.StatusNotImplemented)
-		w.Write([]byte("Not implemented"))
-	}
-}
-
-//Sends waypoints to the plane
-type missionHandlerStart struct {
-	server           *Server
-	waypointChan     chan *pp.Path
-	missionStartChan chan bool
-}
-
-func (m missionHandlerStart) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-	logRequestInfo(r)
-	switch r.Method {
-	case "POST":
+		targets := api.Group("/targets")
 		{
-			// sends waypoints from path planning to a waypoint channel (which will be transmitted to the plane)
-			m.waypointChan <- m.server.path
-			m.missionStartChan <- true
-			message := fmt.Sprintf("Attempting to send %d waypoints to plane", len(m.server.path.GetPath()))
-			w.Write([]byte(message))
-			Log.Info(message)
+			targets.POST("/unclassified", server.postOBCTargets())
+			targets.POST("/classified", server.postCVSResults())
+
+			targets.GET("/unclassified", server.getOBCTargets())
+			targets.GET("/classified", server.getStoredCVSResults())
+
+			targets.PUT("/classified", server.updateCVSResults())
 		}
-	default:
+
+		plane := api.Group("/plane")
 		{
-			w.WriteHeader(http.StatusNotImplemented)
-			w.Write([]byte("Not implemented"))
+			plane.POST("/airdrop", server.uploadDropOrder())
+			plane.GET("/airdrop", server.getDropOrder())
+			plane.PATCH("/airdrop", server.updateDropOrder())
+
+			plane.GET("/telemetry/history", server.getTelemetryHistory())
+			plane.GET("/telemetry", server.getTelemetry())
+
+			plane.GET("/position/history", server.getPositionHistory())
+			plane.GET("/position", server.getPosition())
+
+			plane.GET("/voltage", server.getBatteryVoltages())
 		}
-	}
-}
 
-type pathCacherHandler struct {
-	server *Server
-}
-
-func (m pathCacherHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-	logRequestInfo(r)
-	switch r.Method {
-	case "GET":
+		mission := api.Group("/mission")
 		{
-			if m.server.path == nil {
-				errMsg := "Path has not been initialized yet"
-				Log.Error(errMsg)
-				w.Write([]byte(errMsg))
-				break
-			}
-			data, err := json.Marshal(m.server.path.Waypoints)
-			if err != nil {
-				Log.Error(err)
-				w.Write([]byte(err.Error()))
-			}
-			w.Write(data)
+			/*change field to flight */
+			mission.GET("/bounds/field", server.getFieldBounds())
+			mission.POST("/bounds/field", server.uploadFieldBounds())
+
+			mission.GET("/bounds/airdrop", server.getAirdropBounds())
+			mission.POST("/bounds/airdrop", server.uploadAirDropBounds())
+
+			mission.GET("/path/initial/new", server.generateNewInitialPath())
+			mission.GET("/path/initial", server.getCurrentInitialPath())
+			mission.POST("/path/initial", server.postInitialPath())
+
+			mission.POST("/waypoints", server.handleInitialWaypoints())
+			mission.GET("/waypoints", server.getInitialWaypoints())
+
+			mission.POST("/start", server.startMission())
+
+			mission.POST("/time", server.startMissionTimer())
+			mission.GET("/time", server.getTimeElapsed())
+
+			mission.GET("/state", server.getState())
+			mission.POST("/state", server.changeState())
+			mission.GET("/state/time", server.getStateStartTime())
+			mission.GET("/state/history", server.getStateHistory())
+
+			mission.GET("/camera/status", server.getCameraStatus())
+			mission.GET("/camera/mock/status", server.getMockCameraStatus())
+
+			mission.POST("/airdrop/manual/drop", server.manualBottleDrop())
+			mission.POST("/airdrop/manual/swap", server.manualBottleSwap())
 		}
-	default:
+
+		mavlink := api.Group("/mavlink")
 		{
-			w.WriteHeader(http.StatusNotImplemented)
-			w.Write([]byte("Not implemented"))
+			mavlink.GET("/endpoints", server.getMavlinkEndpoints())
+			mavlink.PUT("/endpoints", server.putMavlinkEndpoints())
 		}
-	}
-}
-
-// Handles uploading and retreiving the home position of the plane
-type planeHomeHandler struct {
-	server *Server
-}
-
-func (p *planeHomeHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-	logRequestInfo(r)
-
-	switch r.Method {
-	case "GET":
-		if p.server.homePosition == nil {
-			w.WriteHeader(http.StatusNotFound)
-			w.Write([]byte("No home position has been set."))
-			Log.Info("No home position has been set.")
-		} else {
-			jsonData, _ := json.Marshal(&p.server.homePosition)
-			w.Write(jsonData)
-			Log.Info("Successfully returned set home position.")
-		}
-		break
-	case "POST":
-		msgBody, _ := ioutil.ReadAll(r.Body)
-		var homePos ic.Position
-		err := json.Unmarshal(msgBody, &homePos)
-		if err != nil {
-			w.WriteHeader(http.StatusBadRequest)
-			w.Write([]byte(fmt.Sprintf("Error parsing home position: %s", err.Error())))
-			Log.Errorf("Unable to parse home position: %s", err.Error())
-			break
-		}
-
-		if homePos.Latitude == nil || homePos.Longitude == nil {
-			w.WriteHeader(http.StatusBadRequest)
-			w.Write([]byte("Error: Latitude and/or longitude not properly set."))
-			Log.Errorf("Latitude and/or longitude not properly set.")
-			break
-		}
-
-		p.server.homePosition = &homePos
-		w.Write([]byte("Successfully updated home position."))
-		Log.Info("Successfully updated home position.")
-		break
-	}
-}
-
-// Handles GET requests that ask for our Plane's telemetry data
-type planeTelemHandler struct {
-	server *Server
-}
-
-func (t *planeTelemHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-	logRequestInfo(r)
-	switch r.Method {
-	case "GET":
-		if t.server.telemetry == nil {
-			w.WriteHeader(http.StatusNotFound)
-			w.Write([]byte("No telemetry found. Is the plane flying?"))
-			Log.Error("No telemetry found. Is the plane flying?")
-			break
-		}
-
-		w.Write(t.server.telemetry)
-		Log.Info("Successfully retrieved our plane's telemetry.")
-	default:
-		w.WriteHeader(http.StatusNotImplemented)
-		w.Write([]byte("Not Implemented"))
-	}
-}
-
-// Handles GET requests that ask for our Plane's telemetry data
-type planeTelemetryHandler struct {
-	server *Server
-	uri    string
-	token  string
-	org    string
-	bucket string
-}
-
-func (t *planeTelemetryHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-	logRequestInfo(r)
-	switch r.Method {
-	case "GET":
-		client := influxdb2.NewClient(t.uri, t.token)
-		queryAPI := client.QueryAPI(t.org)
-		id := r.URL.Query().Get("id")
-		fieldsSeparatedByCommas := r.URL.Query().Get("field")
-		if id == "" {
-			w.WriteHeader(http.StatusBadRequest)
-			w.Write([]byte("Missing id parameter"))
-			break
-		}
-		if fieldsSeparatedByCommas == "" {
-			w.WriteHeader(http.StatusBadRequest)
-			w.Write([]byte("Missing field parameter"))
-			break
-		}
-		// split up the field param by comma
-		// fields is an array where each index is a value we want to get from the database
-		fields := strings.Split(fieldsSeparatedByCommas, ",")
-		// each fieldString is a query string with one of the fields from fields
-		queryStrings := []string{}
-		for _, f := range fields {
-			Log.Infof("current f: %s", f)
-			queryStrings = append(queryStrings,
-				fmt.Sprintf(`from(bucket:"%s") |> range(start: -5m) |> tail(n: 1, offset: 0) |> filter(fn: (r) => r.ID == "%s") |> filter(fn: (r) => r._field == "%s")`,
-					t.bucket, id, f))
-		}
-		// Go through the query strings we made and put them in this results slice
-		// results := []influxdb2.QueryTableResult{}
-		var results []string
-		for _, queryString := range queryStrings {
-			result, err := queryAPI.Query(context.Background(), queryString)
-			if err != nil {
-				w.WriteHeader(http.StatusInternalServerError)
-				w.Write([]byte(fmt.Sprintf("Error Querying InfluxDB: %s", err)))
-				return
-			} else {
-				if result.Next() {
-					val := fmt.Sprint(result.Record().Value())
-					results = append(results, val)
-				} else {
-					w.WriteHeader(http.StatusBadRequest)
-					w.Write([]byte(fmt.Sprintf("Requested telemetry with query %s not found in InfluxDB. Check the id and field in the Mavlink documentation at http://mavlink.io/en/messages/common.html", queryString)))
-					return
-				}
-			}
-		}
-		jsonMap := make(map[string]interface{})
-		for i, field := range fields {
-			jsonMap[field] = results[i]
-		}
-		jsonStr, _ := json.Marshal(jsonMap)
-		w.Write([]byte(jsonStr))
-
-		client.Close()
-	default:
-		w.WriteHeader(http.StatusNotImplemented)
-		w.Write([]byte("Not Implemented"))
-	}
-}
-
-// Handles GET requests that ask for Team Status information
-type interopTeamHandler struct {
-	server *Server
-}
-
-func (t *interopTeamHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-	logRequestInfo(r)
-
-	switch r.Method {
-	case "GET":
-		if t.server.client == nil {
-			w.WriteHeader(http.StatusInternalServerError)
-			w.Write([]byte("Interop connection not established"))
-			Log.Errorf("Unable to retrieve team data from Interop because connection to Interop not established")
-			return
-		}
-
-		// Make the GET request to the Interop Server
-		teams, err := t.server.client.GetTeams()
-		if err.Get {
-			w.WriteHeader(err.Status)
-			w.Write(err.Message)
-		} else {
-			w.Write(teams)
-		}
-	default:
-		w.WriteHeader(http.StatusNotImplemented)
-		w.Write([]byte("Not Implemented"))
-	}
-}
-
-// Handles GET requests that ask for the mission parameters
-type interopMissionHandler struct {
-	server *Server
-}
-
-func (m *interopMissionHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-	logRequestInfo(r)
-
-	switch r.Method {
-	case "GET":
-		// Make the GET request to the interop server
-		if m.server.client == nil {
-			w.WriteHeader(http.StatusInternalServerError)
-			w.Write([]byte("Interop connection not established"))
-			Log.Errorf("Unable to retrieve mission data from Interop because connection to Interop not established")
-			return
-		}
-
-		mission, err := m.server.client.GetMission(m.server.missionID.ID)
-		if err.Get {
-			w.WriteHeader(err.Status)
-			w.Write(err.Message)
-			Log.Errorf("Unable to retrieve mission data from Interop: %s", err.Message)
-		} else {
-			w.Write(mission)
-			Log.Info("Successfully retrieved mission from Interop.")
-		}
-
-	default:
-		w.WriteHeader(http.StatusNotImplemented)
-		w.Write([]byte("Not Implemented"))
-	}
-}
-
-// Handles POST requests to the server that upload telemetry data
-type interopTelemHandler struct {
-	server *Server
-}
-
-func (t *interopTelemHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-	logRequestInfo(r)
-
-	switch r.Method {
-	case "GET":
-		if t.server.client == nil {
-			w.WriteHeader(http.StatusInternalServerError)
-			w.Write([]byte("Interop connection not established"))
-			Log.Errorf("Unable to retrieve telemetry data from Interop because connection to Interop not established")
-			return
-		}
-
-		// We want to parse the teams data to find all of the telemetry of the other planes
-		teamsData, _ := t.server.client.GetTeams()
-		var teamsList []ic.TeamStatus
-		json.Unmarshal(teamsData, &teamsList)
-
-		// We have a list of TeamStatuses in teamsList, now convert to a list of
-		// Telemetry, and return that back into json
-		var telemList []ic.Telemetry
-		for i := 0; i < len(teamsList); i++ {
-			team := &teamsList[i]
-
-			// We don't want to get our own telemety or telemetry from planes
-			// not in the air, so filter out those
-			if team.GetTeam().GetUsername() != t.server.client.GetUsername() && team.GetInAir() {
-				// To prevent a crash if a team has taken off but not uploaded any telemetry
-				if team.GetTelemetry() != nil {
-					telemList = append(telemList, *team.GetTelemetry())
-				}
-			}
-		}
-
-		// Now telemlist should have all the other teams telemetry, so lets turn it back into
-		// a []byte
-		telemData, _ := json.Marshal(telemList)
-		if len(telemList) > 0 {
-			w.Write(telemData)
-			Log.Infof("Successfully retrieved telemetry data from %d other team(s) flying right now", len(telemList))
-		} else {
-			w.Write([]byte("There are no other teams in the air transmitting telemetry."))
-			Log.Infof("There are no other teams in the air transmitting telemtry.")
-		}
-
-	case "POST":
-		if t.server.client == nil {
-			w.WriteHeader(http.StatusInternalServerError)
-			w.Write([]byte("Interop connection not established"))
-			Log.Errorf("Unable to post telemetry data to Interop because connection to Interop not established")
-			return
-		}
-
-		telemData, _ := ioutil.ReadAll(r.Body)
-		// Make the POST request to the interop server
-		err := t.server.client.PostTelemetry(telemData)
-		if err.Post {
-			w.WriteHeader(err.Status)
-			w.Write(err.Message)
-			Log.Errorf("Unable to post telemetry data to Interop: %s", err.Message)
-		} else {
-			w.Write([]byte("Telemetry successfully uploaded"))
-			Log.Info("Successfully uploaded telemetry data to Interop.")
-		}
-	default:
-		w.WriteHeader(http.StatusNotImplemented)
-		w.Write([]byte("Not Implemented"))
-	}
-}
-
-// Handles all requests related to singular odlc
-// e.g. /hub/interop/odlc/
-type interopOdlcHandler struct {
-	server *Server
-}
-
-func (o *interopOdlcHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-	logRequestInfo(r)
-	if o.server.client == nil {
-		w.WriteHeader(http.StatusInternalServerError)
-		w.Write([]byte("Interop connection not established"))
-		Log.Errorf("Unable to get odlc data from Interop because connection to Interop not established")
-		return
-	}
-
-	splitURI := strings.Split(r.URL.Path, "/")
-	odlcID, _ := strconv.Atoi(splitURI[len(splitURI)-1])
-
-	switch r.Method {
-	case "GET":
-		odlcData, intErr := o.server.client.GetODLC(odlcID)
-		if intErr.Get {
-			w.WriteHeader(intErr.Status)
-			w.Write(intErr.Message)
-			Log.Errorf("Unable to retrieve ODLC %d from Interop: %s", odlcID, intErr.Message)
-		} else {
-			// Everything is OK!
-			// This Write statment corresponds to a successful GET request in the format:
-			// GET /interop/odlcs/X where X is a valid integer
-			w.Write(odlcData)
-			Log.Infof("Successfully retrieved ODLC %d from Interop", odlcID)
-		}
-	case "PUT":
-		odlcData, _ := ioutil.ReadAll(r.Body)
-		updatedOdlc, err := o.server.client.PutODLC(odlcID, odlcData)
-		if err.Put {
-			w.WriteHeader(err.Status)
-			w.Write(err.Message)
-			Log.Errorf("Unable to update ODLC %d on Interop: %s", odlcID, err.Message)
-		} else {
-			// This Write statement corresponds to a successful PUT request in the format:
-			// PUT /interop/odlcs/X where X is a valid integer
-			w.Write(updatedOdlc)
-			Log.Infof("Successfully updated ODLC %d on Interop", odlcID)
-		}
-	case "DELETE":
-		err := o.server.client.DeleteODLC(odlcID)
-		if err.Delete {
-			w.WriteHeader(err.Status)
-			w.Write(err.Message)
-			Log.Errorf("Unable to delete ODLC %d on Interop: %s", odlcID, err.Message)
-		} else {
-			// This Write statement corresponds to a successful DELETE request in the format:
-			// DELETE /interop/odlcs/X where X is a valid integer
-			w.Write([]byte(fmt.Sprintf("Successfully deleted odlc %d", odlcID)))
-			Log.Infof("Successfuly deleted ODLC %d on Interop", odlcID)
-		}
-	default:
-		w.WriteHeader(http.StatusNotImplemented)
-		w.Write([]byte("Not Implemented"))
-	}
-}
-
-//Handles requests for multiple odlcs
-type interopOdlcsHandler struct {
-	server *Server
-}
-
-func (o *interopOdlcsHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-	logRequestInfo(r)
-	if o.server.client == nil {
-		w.WriteHeader(http.StatusInternalServerError)
-		w.Write([]byte("Interop connection not established"))
-		Log.Errorf("Unable to get odlc data from Interop because connection to Interop not established")
-		return
-	}
-
-	switch r.Method {
-	case "GET":
-		odlcsData, intErr := o.server.client.GetODLCs(-1)
-		if intErr.Get {
-			w.WriteHeader(intErr.Status)
-			w.Write(intErr.Message)
-			Log.Errorf("Unable to retrieve ODLCs from Interop: %s", intErr.Message)
-		} else {
-			// Everything is OK!
-			// This Write statement corresponds to a successful GET request in the format:
-			// GET /interop/odlcs/
-			w.Write(odlcsData)
-			Log.Infof("Successfully retrieved ODLCs from Interop")
-		}
-	case "POST":
-		odlcData, _ := ioutil.ReadAll(r.Body)
-		// Make the POST request to the interop server
-		updatedODLC, err := o.server.client.PostODLC(odlcData)
-		if err.Post {
-			w.WriteHeader(err.Status)
-			w.Write(err.Message)
-			Log.Errorf("Unable to upload ODLC to Interop: %s", err.Message)
-		} else {
-			// This Write statement corresponds to a successful POST request in the format:
-			// POST /interop/odlcs
-			w.Write(updatedODLC)
-			Log.Infof("Successfully uploaded ODLC to Interop")
-		}
-	default:
-		w.WriteHeader(http.StatusNotImplemented)
-		w.Write([]byte("Not Implemented"))
-	}
-}
-
-//Handles requests for odlc images
-type interopOdlcImageHandler struct {
-	server *Server
-}
-
-func (o *interopOdlcImageHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-	logRequestInfo(r)
-	if o.server.client == nil {
-		w.WriteHeader(http.StatusInternalServerError)
-		w.Write([]byte("Interop connection not established"))
-		Log.Errorf("Unable to get odlc data from Interop because connection to Interop not established")
-		return
-	}
-
-	splitURI := strings.Split(r.URL.Path, "/")
-	missionID, _ := strconv.Atoi(splitURI[5])
-
-	switch r.Method {
-	case "GET":
-		image, err := o.server.client.GetODLCImage(missionID)
-		if err.Get {
-			w.WriteHeader(err.Status)
-			w.Write(err.Message)
-			Log.Errorf("Unable to get ODLC image from Interop: %s", err.Message)
-		} else {
-			// This Write statement corresponds to a successful request in the format
-			// GET /interop/odlcs/X/image
-			w.Write(image)
-			Log.Info("Successfully retrieved ODLC image from Interop.")
-		}
-	case "PUT":
-		image, _ := ioutil.ReadAll(r.Body)
-		err := o.server.client.PutODLCImage(missionID, image)
-		if err.Put {
-			w.WriteHeader(err.Status)
-			w.Write(err.Message)
-			Log.Errorf("Unable to update ODLC image on Interop: %s", err.Message)
-		} else {
-			// This Write statement corresponds to a successful request in the format
-			// PUT /interop/odlcs/X/image
-			w.Write([]byte(fmt.Sprintf("Successfully uploaded odlc image for odlc %d", missionID)))
-			Log.Infof("Successfully uploaded ODLC image for ODLC %d", missionID)
-		}
-	case "DELETE":
-		err := o.server.client.DeleteODLCImage(missionID)
-		if err.Delete {
-			w.WriteHeader(err.Status)
-			w.Write(err.Message)
-			Log.Errorf("Unable to update ODLC image on Interop: %s", err.Message)
-		} else {
-			// This Write statement corresponds to a successful request in the format
-			// DELETE /interop/odlcs/X/image
-			w.Write([]byte(fmt.Sprintf("Successfully deleted ODLC image for ODLC %d", missionID)))
-			Log.Infof("Successfully deleted ODLC image for ODLC %d", missionID)
-		}
-	default:
-		w.WriteHeader(http.StatusNotImplemented)
-		w.Write([]byte("Not implemented"))
 	}
 }
 
 /*
-Use path planning client to send the static mission data to path planning when Hub receives a POST request
-at /hub/pathplanning/initialize with the following request body { "id": [mission_id] }.
-Then should use the path planning client to forward the static mission data to path planning
+Initializes all http request routes (tells the server which handler functions to call when a certain route is requested).
+
+General route format is "/place/thing".
 */
+func (server *Server) SetupRouter() *gin.Engine {
+	router := gin.Default()
+	router.Use(CORSMiddleware())
 
-// Handles requests from the jetson that include the cropped/salienced image
-type CVCroppedHandler struct {
-	server *Server
-	uri    string
-	token  string
-	org    string
-	bucket string
+	server.initBackend(router)
+	server.initFrontend(router)
+
+	return router
 }
 
-// TODO: extremely scuffed code that needs refactoring. need to make a CVS and
-// OBC client
-func (h CVCroppedHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-	logRequestInfo(r)
+// New will initialize a server struct and populate fields with their initial state
+func New(influxdbClient *influxdb.Client, mavlinkClient *mav.Client, obcClient *obc.Client) Server {
+	server := Server{}
 
-	if h.server.cvData == nil {
-		w.WriteHeader(http.StatusInternalServerError)
-		w.Write([]byte("Computer Vision Server connection not established"))
-		Log.Errorf("Unable to receive data from CVS; connection not established")
-		return
+	server.influxDBClient = influxdbClient
+	server.mavlinkClient = mavlinkClient
+	server.obcClient = obcClient
+
+	server.AirDropBounds = nil
+	server.FlightBounds = nil
+
+	return server
+}
+
+/*
+Starts the server on localhost:5000. Make sure nothing else runs on port 5000 if you want the plane to fly.
+*/
+func (server *Server) Start() {
+	router := server.SetupRouter()
+	server.Manager = manager.NewManager()
+
+	err := router.Run(":5000")
+	if err != nil {
+		// TODO: decide if we want to make this a Log.Fatal and have Hub shutdown
+		Log.Errorf("Gin Engine crashed with the following error: %s", err)
 	}
+}
 
-	switch r.Method {
-	case "POST":
-		var t cv.UnclassifiedODLC
-		bodyData, _ := ioutil.ReadAll(r.Body)
-		err := json.Unmarshal(bodyData, &t)
-		if err != nil {
-			Log.Fatal(err)
+/*
+User testing all of hubs connections. Returns JSON of all the connection statuses.
+TODO: Actually test the connections instead of just returning True.
+*/
+func (server *Server) testConnections() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		c.JSON(http.StatusOK, gin.H{
+			"cvs":             true,
+			"plane_obc":       true,
+			"plane_mavlink":   server.mavlinkClient.IsConnectedToPlane(),
+			"antenna_tracker": server.mavlinkClient.IsConnectedToAntennaTracker()})
+	}
+}
+
+// getTelemetryHistory gets the telemetry from a certain point in time until now.
+// Returns a list of messages. First element in the list is the one at the earliest time.
+// Each message JSON has a "_time" key that is the timestamp of that message.
+// Use query params to specify the message id or name, time range and message fields.
+//
+// Example URL: localhost:5000/api/plane/telemetry?id=33&range=5&fields=alt,hdg
+//
+// Note that only one of ID or name is required. If both are provided, it will
+// default to lookup the ID and ignore the name.
+//
+// URL Params:
+//   - id will be the mavlink message ID. Example: 33
+//   - name will be the mavlink message name. Example: "GLOBAL_POSITION_INT"
+//     A full list of mavlink message names and IDs can be found here
+//     http://mavlink.io/en/messages/common.html
+//   - range is the number of minutes to look back in the past for a message. Can be a floating point number.
+//   - fields are the fields of the mavlink message to return. If none are specified then
+//     all the fields are returned. The fields are separated by commas. Example: "alt,hdg"
+func (server *Server) getTelemetryHistory() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		msgID := c.Query("id")
+		msgName := c.Query("name")
+		timeRange := c.Query("range")
+		fieldsSeparatedByCommas := c.Query("fields")
+
+		if timeRange == "" {
+			c.String(http.StatusBadRequest, "No time range provided")
+			return
 		}
 
-		client := influxdb2.NewClient(h.uri, h.token)
-		queryAPI := client.QueryAPI(h.org)
-		Log.Info("Queryapi", queryAPI)
+		timeRangeFloat, err := strconv.ParseFloat(timeRange, 32)
+		if err != nil {
+			c.String(http.StatusBadRequest, "Non-numerical range provided")
+			return
+		}
 
-		telem := make(map[string]float64)
-		fields := []string{"lat", "lon", "relative_alt", "hdg"}
-		for _, field := range fields {
-			planeLatQuery := fmt.Sprintf(`from(bucket:"%s") |> range(start: %s) |> filter(fn: (r) => r.ID == "33") |> filter(fn: (r) => r._field == "%s") |> first()`, h.bucket, t.Timestamp, field)
-			result, err := queryAPI.Query(context.Background(), planeLatQuery)
+		fields := []string{}
+		if fieldsSeparatedByCommas != "" {
+			fields = strings.Split(fieldsSeparatedByCommas, ",")
+		}
+
+		if msgID != "" {
+			msgIDInt, err := strconv.Atoi(msgID)
 			if err != nil {
-				Log.Error(err)
-			}
-			if result != nil && result.Next() {
-				var ok bool
-				telem[field], ok = result.Record().Value().(float64)
-				if !ok {
-					telem[field] = 0
-					Log.Errorf("Could not parse %s from InfluxDB", field)
-				}
+				c.String(http.StatusBadRequest, "Non-numerical message ID requested")
 			} else {
-				Log.Debugf("Could not find %s field in InfluxDB", field)
+				data, err := server.influxDBClient.QueryMsgIDAndFields(uint32(msgIDInt), time.Duration(timeRangeFloat)*time.Minute, fields...)
+				if err != nil {
+					// TODO: have other types of errors (id does not exist for example)
+					c.String(http.StatusInternalServerError, "Error processing database query. Reason: %s", err)
+					return
+				}
+
+				c.JSON(http.StatusOK, data)
+				return
 			}
 		}
-		t.PlaneLat = telem[fields[0]] / 1e7
-		t.PlaneLon = telem[fields[1]] / 1e7
-		t.PlaneAlt = telem[fields[2]] / 1000
-		t.PlaneHead = telem[fields[3]] / 100
 
-		data, err := json.Marshal(t)
-		if err != nil {
-			Log.Error(err)
-			w.Write([]byte(err.Error()))
-			break
-		}
-		Log.Info("this print statement v")
-		Log.Info(t.PlaneLat)
-		Log.Info(t.PlaneLon)
-		Log.Info(t.PlaneAlt)
-		Log.Info(t.PlaneHead)
-		//Log.Info(string(data))
-		// resp, err := http.Post("http://localhost:5040/upload", "application/json", bytes.NewBuffer(data))
-		httpClient := ut.NewClient("172.17.0.1:5040", 30) // TODO: change this to be env variable
-		// TODO: also make a CVS client (maybe OBC client as well)
-		_, httpErr := httpClient.Post("/upload", bytes.NewBuffer(data))
-		if httpErr.Post {
-			Log.Error("failed to post cropped target to CVS: ", string(httpErr.Message))
+		if msgName != "" {
+			data, err := server.influxDBClient.QueryMsgNameAndFields(msgName, time.Duration(timeRangeFloat)*time.Minute, fields...)
+			if err != nil {
+				// TODO: have other types of errors (name does not exist for example)
+				c.String(http.StatusInternalServerError, "Error processing database query. Reason: %s", err)
+				return
+			}
+
+			c.JSON(http.StatusOK, data)
 			return
 		}
 
-	default:
-		w.WriteHeader(http.StatusNotImplemented)
-		w.Write([]byte("Not implemented"))
+		c.String(http.StatusBadRequest, "No message name or ID provided")
 	}
 }
 
-// Handles requests from the computer vision server with the predicted results
-type CVResultHandler struct {
-	server *Server
+// getTelemetry gets the latest telemetry.
+// Use query params to specify the message id, name and message fields.
+//
+// Example URL: localhost:5000/api/plane/telemetry?id=33&range=5&fields=alt,hdg
+//
+// Note that only one of ID or name is required. If both are provided, it will
+// default to lookup the ID and ignore the name.
+//
+// URL Params:
+//   - id will be the mavlink message ID. Example: 33
+//   - name will be the mavlink message name. Example: "GLOBAL_POSITION_INT"
+//     A full list of mavlink message names and IDs can be found here
+//     http://mavlink.io/en/messages/common.html
+//   - fields are the fields of the mavlink message to return. If none are specified then
+//     all the fields are returned. The fields are separated by commas. Example: "alt,hdg".
+func (server *Server) getTelemetry() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		msgID := c.Query("id")
+		msgName := c.Query("name")
+		fieldsSeparatedByCommas := c.Query("fields")
+
+		fields := []string{}
+		if fieldsSeparatedByCommas != "" {
+			fields = strings.Split(fieldsSeparatedByCommas, ",")
+		}
+
+		if msgID != "" {
+			msgIDInt, err := strconv.Atoi(msgID)
+			if err != nil {
+				c.String(http.StatusBadRequest, "Non-numerical message ID requested")
+				return
+			}
+
+			data, err := server.influxDBClient.QueryMsgIDAndFields(uint32(msgIDInt), 0, fields...)
+			if err != nil {
+				// TODO: have other types of errors (id does not exist for example)
+				c.String(http.StatusInternalServerError, "Error processing database query. Reason: %s", err)
+				return
+			}
+
+			if len(data) == 0 {
+				c.String(http.StatusNotFound, "No telemetry found")
+				return
+			}
+
+			// only return the 0th index since data will always return a list with a single element if timeRange is 0
+			c.JSON(http.StatusOK, data[0])
+			return
+		}
+
+		if msgName != "" {
+			data, err := server.influxDBClient.QueryMsgNameAndFields(msgName, 0, fields...)
+			if err != nil {
+				// TODO: have other types of errors (name does not exist for example)
+				c.String(http.StatusInternalServerError, "Error processing database query. Reason: %s", err)
+				return
+			}
+
+			if len(data) == 0 {
+				c.String(http.StatusNotFound, "No telemetry found")
+				return
+			}
+
+			c.JSON(http.StatusOK, data[0])
+			return
+		}
+
+		c.String(http.StatusBadRequest, "No message name or ID provided")
+	}
 }
 
-func (h CVResultHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-	logRequestInfo(r)
+// getPositionHistory gets the plane position from a certain point in the past.
+// Returns a list of position telemetry.
+//
+// Matches format of GLOBAL_POSITION_INT mavlink message.
+// https://mavlink.io/en/messages/common.html#GLOBAL_POSITION_INT
+//
+// URL Params:
+//   - range is the number of minutes to look back in the past for a message. Can be a floating point number.
+func (server *Server) getPositionHistory() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		timeRange := c.Query("range")
 
-	if h.server.cvData == nil {
-		w.WriteHeader(http.StatusInternalServerError)
-		w.Write([]byte("Computer Vision Server connection not established"))
-		Log.Errorf("Unable to receive data from CVS; connection not established")
-		return
+		timeRangeFloat, err := strconv.ParseFloat(timeRange, 32)
+		if err != nil {
+			c.String(http.StatusBadRequest, "Non-numerical range provided")
+			return
+		}
+
+		data, err := server.influxDBClient.QueryMsgID(33, time.Duration(timeRangeFloat)*time.Minute)
+		if err != nil {
+			// TODO: have other types of errors (id does not exist for example)
+			c.String(http.StatusInternalServerError, "Error processing database query. Reason: %s", err)
+			return
+		}
+
+		c.JSON(http.StatusOK, data)
+	}
+}
+
+// getPosition gets the latest plane position.
+//
+// Matches format of GLOBAL_POSITION_INT mavlink message.
+// https://mavlink.io/en/messages/common.html#GLOBAL_POSITION_INT
+func (server *Server) getPosition() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		data, err := server.influxDBClient.QueryMsgID(33, 0)
+		if err != nil {
+			// TODO: have other types of errors (id does not exist for example)
+			c.String(http.StatusInternalServerError, "Error processing database query. Reason: %s", err)
+			return
+		}
+
+		if len(data) == 0 {
+			c.String(http.StatusNotFound, "No telemetry found")
+			return
+		}
+
+		c.JSON(http.StatusOK, data[0])
+	}
+}
+
+// getBatteryVoltages retrieves the latest voltage information from the mavlink client
+func (server *Server) getBatteryVoltages() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		c.JSON(http.StatusOK, server.mavlinkClient.LatestBatteryInfo)
+	}
+}
+
+// getMavlinkEndpoints responds with the mavlink endpoints that Hub is currently
+// communicating with. This includes the plane itself and devices that are receiving
+// mavlink messages through Hub's mavlink router.
+//
+// Will return an error if the plane endpoint has not been created yet.
+func (server *Server) getMavlinkEndpoints() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		planeEndpoint, err := server.mavlinkClient.GetPlaneEndpoint()
+		if err != nil {
+			c.String(http.StatusNotFound, err.Error())
+			return
+		}
+		routerEndpoints := server.mavlinkClient.GetRouterEndpoints()
+		endpointData := mav.EndpointData{
+			Plane:  planeEndpoint,
+			Router: routerEndpoints,
+		}
+		c.JSON(http.StatusOK, endpointData)
+	}
+}
+
+// putMavlinkEndpoints will update the plane and router mavlink endpoints.
+//
+// The JSON body should match the mav.EndpointData struct.
+//
+// Example body:
+//
+//	{
+//			"plane": "serial:/dev/ttyUSB0",
+//			"router": [
+//						"udp:192.168.1.7:14551",
+//						"tcp:localhost:14550"
+//					  ]
+//	}
+func (server *Server) putMavlinkEndpoints() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		endpointData := mav.EndpointData{}
+		err := c.BindJSON(&endpointData)
+		if err != nil {
+			c.String(http.StatusBadRequest, err.Error())
+			return
+		}
+
+		server.mavlinkClient.UpdateEndpoints(endpointData.Plane, endpointData.Router)
+		c.String(http.StatusOK, "Updated mavlink endpoints")
+	}
+}
+
+/*
+User (plane/jetson) sends cropped image, bounding box, and other plane telemetry, and it is saved in the server struct.
+*/
+func (server *Server) postOBCTargets() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		unclassifiedODLCData := []cvs.UnclassifiedODLC{}
+		err := c.BindJSON(&unclassifiedODLCData)
+
+		if err == nil {
+			server.UnclassifiedTargets = append(server.UnclassifiedTargets, unclassifiedODLCData...)
+			c.String(http.StatusOK, "Accepted ODLC data")
+			return
+		}
+		c.String(http.StatusBadRequest, err.Error())
+	}
+}
+
+/*
+Returns the unclassified targets that are currently stored in Hub,
+or an error if they have not been posted yet.
+*/
+func (server *Server) getOBCTargets() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		if server.UnclassifiedTargets == nil {
+			c.String(http.StatusBadRequest, "ERROR: Unclassified Targets have not been posted yet")
+		} else {
+			c.JSON(http.StatusOK, server.UnclassifiedTargets)
+		}
+	}
+}
+
+/*
+Returns an integer representing the Unix time of when startMissionTimer() was called.
+This is intended to be passed to Houston, which will then convert it to the time since the mission started.
+*/
+func (server *Server) getTimeElapsed() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		// if time hasn't been initialized yet, throw error
+		if server.MissionTime == 0 {
+			c.String(http.StatusBadRequest, "ERROR: time hasn't been initalized yet") // not sure if there's a built-in error message to use here
+		} else {
+			c.String(http.StatusOK, fmt.Sprint(server.MissionTime))
+		}
+	}
+}
+
+/*
+Starts a timer when the mission begins, in order to keep track of how long the mission has gone on.
+*/
+func (server *Server) startMissionTimer() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		server.MissionTime = time.Now().Unix()
+		c.String(http.StatusOK, "Mission timer successfully started!")
+	}
+}
+
+/*
+Query Hub for the mission's current state
+*/
+func (server *Server) getState() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		c.String(http.StatusOK, server.Manager.State.String())
+	}
+}
+
+/*
+Request a change to the mission's state.
+
+If the state is changing to "AIRDROP APPROACH" (signaling the start of the airdrop sequence),
+it will post the classified ODLCs to the OBC using the client function.
+
+The request will error with code 409 CONFLICT if it is an invalid state change.
+*/
+func (server *Server) changeState() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		stateJSON := manager.StateJSON{}
+		err := c.BindJSON(&stateJSON)
+
+		if err != nil {
+			c.String(http.StatusBadRequest, err.Error())
+		}
+
+		prevState := server.Manager.State
+		state := stateJSON.ToEnum()
+
+		if server.Manager.ChangeState(state) {
+			// if state.String() == "AIRDROP APPROACH" {
+			//TODO: FINISH THIS WHEN THE OBC SERVER HAS A ROUTE FOR UPLOADING CLASSIFIED ODLCS
+			// }
+			c.String(http.StatusOK, fmt.Sprintf("Successful state change: %s to %s.", prevState, state))
+		} else {
+			c.String(http.StatusConflict, fmt.Sprintf("Invalid state change: %s to %s.", prevState, state))
+		}
+	}
+}
+
+/*
+Returns the starting time of the current state in seconds since epoch
+*/
+func (server *Server) getStateStartTime() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		c.String(http.StatusOK, fmt.Sprintf("%d", server.Manager.GetCurrentStateStartTime()))
+	}
+}
+
+/*
+Returns the list of state changes that have occurred, with timestamps
+*/
+func (server *Server) getStateHistory() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		c.JSON(http.StatusOK, server.Manager.HistoryJSON())
+	}
+}
+
+/*
+User (person manning ground control station) will type in the ODLC info on each water bottle as well as the ordering of each bottle in the plane, and then click a button on Houston to upload it.
+
+Also note that one of the bottles should dropped on a manikin (won't have alphanumeric/color?)
+
+IDEA (to implement in the future): check to make sure the length of the bottle slice is no more than 5 (there should only be 5 bottles uploaded, but if there's some error we might want to be able to upload less).
+*/
+func (server *Server) uploadDropOrder() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		bottleOrdering := airdrop.Bottles{}
+
+		err := c.BindJSON(&bottleOrdering.Bottles)
+
+		if err == nil {
+			server.Bottles = &bottleOrdering
+			c.String(http.StatusOK, "Bottles successfully uploaded!")
+		} else {
+			c.String(http.StatusBadRequest, err.Error())
+		}
+	}
+}
+
+/*
+Returns the information (drop index, which target to drop on, etc.) about each water bottle as it has been entered by the person manning the ground control station.
+*/
+func (server *Server) getDropOrder() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		if server.Bottles == nil {
+			c.String(http.StatusBadRequest, "ERROR: drop order not yet initialized")
+		} else {
+			c.JSON(http.StatusOK, server.Bottles.Bottles)
+		}
+	}
+}
+
+/*
+Updates the information about a single water bottle as it was entered by the person manning the ground control station.
+
+This would be useful if there was a mistake when uploading the original drop ordering, so the information for a single bottle can be updated given an id.
+
+NOTE: this is currently not capable of updating a bottle's dropIndex in case it was entered incorrectly (because identifying the bottle to update is dependent on the dropIndex in the patch request body).
+If we want to change this, one solution would be to have an extra field in the json body for the dropIndex to update, and then in the proceeding struct the user would be able to enter a replacement dropIndex.
+
+If all bottles need to be updated at once, the user should just use the POST request (uploadDropOrder(); will clear everything and overwrite with the post body).
+*/
+func (server *Server) updateDropOrder() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		bottleToUpdate := airdrop.Bottle{}
+		err := c.BindJSON(&bottleToUpdate)
+
+		bottleUpdated := false
+
+		// loop through each of the bottles that are currently uploaded and find the one with the right id, and then overwrite its values
+		for i, bottle := range server.Bottles.Bottles {
+			if bottle.DropIndex == bottleToUpdate.DropIndex {
+				server.Bottles.Bottles[i] = bottleToUpdate
+				bottleUpdated = true
+			}
+		}
+
+		if err != nil {
+			c.String(http.StatusBadRequest, err.Error())
+			return
+		}
+
+		if bottleUpdated {
+			c.String(http.StatusOK, "Bottle %d has been updated!", bottleToUpdate.DropIndex)
+		} else {
+			c.String(http.StatusBadRequest, "ERROR: Bottle %d not found. Make sure the entire drop order has been initialized before updating individual bottles!", bottleToUpdate.DropIndex)
+		}
+	}
+}
+
+/*
+Returns the slice of Coordinates representing the mission boundaries stored in server,
+or an error if it has not been initialized.
+*/
+func (server *Server) getFieldBounds() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		if server.FlightBounds == nil {
+			c.String(http.StatusBadRequest, "ERROR: Flight bounds not yet initialized")
+		} else {
+			c.JSON(http.StatusOK, server.FlightBounds)
+		}
+	}
+}
+
+/*
+Reads in longitude and latitude coordinates for field bounds and uploads to the server
+*/
+func (server *Server) uploadFieldBounds() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		fieldBounds := []pp.Coordinate{}
+		err := c.BindJSON(&fieldBounds)
+
+		if err == nil {
+			server.FlightBounds = fieldBounds
+			body, status := server.uploadMissionIfReady()
+			if status == -1 {
+				c.String(http.StatusOK, "Field Bounds Uploaded. Still need Airdrop Bounds.")
+			} else {
+				c.String(status, fmt.Sprintf("Field Bounds uploaded: Attempt to upload mission: %s", body))
+			}
+		} else {
+			c.String(http.StatusBadRequest, err.Error())
+		}
+	}
+}
+
+func (server *Server) uploadMissionIfReady() (string, int) {
+	if server.FlightBounds != nil && server.AirDropBounds != nil {
+		mission := pp.Mission{
+			FlightBoundaries: server.FlightBounds,
+			SearchBoundaries: server.AirDropBounds,
+		}
+
+		body, status := server.obcClient.PostMission(&mission)
+		return string(body), status
 	}
 
-	switch r.Method {
-	case "POST":
-		imageData, _ := ioutil.ReadAll(r.Body)
+	return "Mission not fully uploaded yet.", -1
+}
 
-		var image cv.ClassifiedODLC
+/*
+Returns the slice of coordinates representing the airdrop zone boundaries that is stored in the server,
+or an error if it has not yet been initialized.
+*/
+func (server *Server) getAirdropBounds() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		if server.AirDropBounds == nil {
+			c.String(http.StatusBadRequest, "ERROR: Airdrop bound not yet initialized")
+		} else {
+			c.JSON(http.StatusOK, server.AirDropBounds)
+		}
+	}
+}
 
-		// unmarshal posted image data and store into image (classified ODLC struct)
-		err := json.Unmarshal(imageData, &image)
+/*
+Reads in longitude and latitude coordinates for airdrop bounds and uploads to the server
+*/
+func (server *Server) uploadAirDropBounds() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		airDropBounds := []pp.Coordinate{}
+		err := c.BindJSON(&airDropBounds)
+
+		if err == nil {
+			server.AirDropBounds = airDropBounds
+			body, status := server.uploadMissionIfReady()
+			if status == -1 {
+				c.String(http.StatusOK, "Airdrop Bounds Uploaded. Still need to upload Flight Boundaries.")
+			} else {
+				c.String(status, fmt.Sprintf("Airdrop Bounds uploaded: Attempt to upload mission: %s", body))
+			}
+		} else {
+			c.String(http.StatusBadRequest, err.Error())
+		}
+	}
+}
+
+/*
+CVS sends results (target coordinates, alphanumeric, shape, color) to Hub and forward target coordinates to RTPP
+*/
+func (server *Server) postCVSResults() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		cvsResults := cvs.ClassifiedODLC{}
+		err := c.BindJSON(&cvsResults)
+
+		if err == nil {
+			c.String(http.StatusOK, "Successfully received CVS results")
+			server.ClassifiedTargets = append(server.ClassifiedTargets, cvsResults)
+			//TODO: forward target coordinates to path planning when OBC2 client is finished
+		} else {
+			c.String(http.StatusBadRequest, err.Error())
+		}
+	}
+}
+
+/*
+Request target data that was posted earlier via postCVSResults
+*/
+func (server *Server) getStoredCVSResults() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		if server.ClassifiedTargets == nil {
+			c.String(http.StatusBadRequest, "ERROR: CVS results have not been posted yet")
+		} else {
+			c.JSON(http.StatusOK, server.ClassifiedTargets)
+		}
+	}
+}
+
+/*
+Returns a JSON containing true if camera is currently taking images; false otherwise.
+
+This is intended for Houston to use.
+*/
+func (server *Server) getCameraStatus() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		status := struct {
+			status bool
+		}{
+			status: server.obcClient.CameraStatus,
+		}
+		c.JSON(http.StatusOK, status)
+	}
+}
+
+/*
+Returns a JSON containing true if the mock camera is currently "taking images"; false otherwise.
+
+This is intended for Houston to use.
+*/
+func (server *Server) getMockCameraStatus() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		status := struct {
+			status bool
+		}{
+			status: server.obcClient.MockCameraStatus,
+		}
+		c.JSON(http.StatusOK, status)
+	}
+}
+
+/*
+Sends request to OBC to generate a new initial Path
+*/
+func (server *Server) generateNewInitialPath() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		path, status := server.obcClient.GenerateNewInitialPath()
+		c.Data(status, "application/json", path)
+	}
+}
+
+/*
+Sends request to OBC for currently uploaded initial path
+*/
+func (server *Server) getCurrentInitialPath() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		path, status := server.obcClient.GetCurrentInitialPath()
+		c.Data(status, "application/json", path)
+	}
+}
+
+/*
+Sends Post request to OBC to upload initial path
+*/
+func (server *Server) postInitialPath() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		path := []pp.Waypoint{}
+		err := c.BindJSON(&path)
 		if err != nil {
-			Log.Error("error unmarshalling image data: ", err)
-			return
+			c.String(http.StatusBadRequest, "Error with json binding %s", err.Error())
 		}
-		h.server.cvData.ClassifiedODLCs = append(h.server.cvData.ClassifiedODLCs, image) // saves unmarshalled data to list of cv images
 
-		var odlc ic.Odlc
+		resp, status := server.obcClient.PostInitialPath(path)
+		c.String(status, string(resp))
+	}
+}
 
-		// assign values to odlc struct
-		var missionID = int32(h.server.missionID.ID)
-		odlc.Mission = &missionID
-		var odlcType = ic.Odlc_STANDARD
-		odlc.Type = &odlcType
-		odlc.Latitude = &image.Latitude
-		odlc.Longitude = &image.Longitude
-		odlc.Orientation = ic.Odlc_Orientation(ic.Odlc_Orientation_value[image.Orientation]).Enum()
-		odlc.Shape = ic.Odlc_Shape(image.Shape).Enum()
-		odlc.Alphanumeric = &image.Char
-		odlc.ShapeColor = ic.Odlc_Color(image.ShapeColor).Enum()
-		odlc.AlphanumericColor = ic.Odlc_Color(image.CharColor).Enum()
-		var autonomous = true
-		odlc.Autonomous = &autonomous
-
-		// have to convert the alphanumeric to uppercase because the PostODLC function returns an error otherwise
-		image.Char = strings.ToUpper(image.Char)
-
-		// marshal values into json
-		jsonStr, _ := json.Marshal(odlc)
-
-		// gets a json of the same odlc returned with the id
-		jsonStr, httpErr := h.server.client.PostODLC(jsonStr)
-		if httpErr.Post {
-			Log.Error("failed to post odlc: ", string(httpErr.Message))
-			return
-		}
-		Log.Info("Posted ODLC to Interop")
-
-		// convert back to odlc
-		err = json.Unmarshal(jsonStr, &odlc)
+/*
+Sends Post request to OBC with the competition waypoints we have to hit (in the initial path)
+*/
+func (server *Server) handleInitialWaypoints() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		wpts := []pp.Waypoint{}
+		err := c.BindJSON(&wpts)
 		if err != nil {
-			Log.Error(err)
-			return
+			c.String(http.StatusBadRequest, "Error with json binding %s", err.Error())
 		}
 
-		// decode base64 data into image
-		data, err := base64.StdEncoding.DecodeString(image.CroppedImageBase64)
+		resp, status := server.obcClient.PostInitialWaypoint(&wpts)
+		c.String(status, string(resp))
+	}
+}
+
+/*
+Gets the initial waypoints that have already been uploaded to the OBC
+*/
+func (server *Server) getInitialWaypoints() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		resp, status := server.obcClient.GetInitialWaypoints()
+		c.Data(status, "application/json", resp)
+	}
+}
+
+/*
+Sends Post request to the OBC to signify that we want to start the autonomous mission
+*/
+func (server *Server) startMission() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		resp, status := server.obcClient.StartMission()
+		c.String(status, string(resp))
+	}
+}
+
+/*
+In case the images were not classified correctly, Houston may need to update the values manually.
+
+This function currently just overwrites the values in server.ClassfiedTargets; we might want to change this to update specific values later.
+*/
+func (server *Server) updateCVSResults() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		updatedResults := []cvs.ClassifiedODLC{}
+		err := c.BindJSON(&updatedResults)
+
 		if err != nil {
-			Log.Error("decoding error: ", err)
-			return
+			c.String(http.StatusBadRequest, err.Error())
+		} else {
+			server.ClassifiedTargets = updatedResults
+			c.String(http.StatusOK, "Successfully updated CVS results!")
 		}
+	}
+}
 
-		// sends odlc with image to server
-		httpErr = h.server.client.PutODLCImage(int(*odlc.Id), data)
-		if httpErr.Put {
-			Log.Error("failed to put odlc: ", httpErr)
-			Log.Error("put fail message: ", string(httpErr.Message))
-			return
+func (server *Server) manualBottleDrop() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		resp, status := server.obcClient.ManualBottleDrop()
+		c.String(status, string(resp))
+	}
+}
+
+func (server *Server) manualBottleSwap() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		bottleSwap := airdrop.BottleSwap{}
+		err := c.BindJSON(&bottleSwap)
+		if err != nil {
+			c.String(http.StatusBadRequest, "Bad JSON formatting: %s", err.Error())
 		}
-		Log.Info("Put ODLC to Interop")
-
-	default:
-		w.WriteHeader(http.StatusNotImplemented)
-		w.Write([]byte("Not implemented"))
+		resp, status := server.obcClient.ManualBottleSwap(bottleSwap)
+		c.String(status, string(resp))
 	}
 }
