@@ -4,46 +4,37 @@ import (
 	"flag"
 	"fmt"
 	"os"
-	"strconv"
 	"strings"
 
 	"github.com/sirupsen/logrus"
-	ic "github.com/tritonuas/hub/internal/interop"
-	mav "github.com/tritonuas/hub/internal/mavlink"
-	pp "github.com/tritonuas/hub/internal/path_plan"
-	hs "github.com/tritonuas/hub/internal/server"
+
+	"github.com/tritonuas/gcs/internal/influxdb"
+	mav "github.com/tritonuas/gcs/internal/mavlink"
+
+	"github.com/tritonuas/gcs/internal/obc"
+	"github.com/tritonuas/gcs/internal/server"
 )
 
 var log = logrus.New()
+
+// Defines globally used variables for ports and IPs and other things.
 var ENVS = map[string]*string{
-	"HUB_ADDR":           flag.String("hub_addr", "5000", "http service hub_address"),
-	"HUB_PATH":           flag.String("hub_path", "/home/mat/gopath/src/github.com/tritonuas/hub", "Path to hub folder"),
-	"INTEROP_IP":         flag.String("interop_ip", "127.0.0.1", "ip of interop computer"),
-	"INTEROP_PORT":       flag.String("interop_port", "8000", "port of interop computer"),
-	"INTEROP_USER":       flag.String("interop_user", "ucsdauvsi", "username on interop computer"),
-	"INTEROP_PASS":       flag.String("interop_pass", "tritons", "password to interop computer"),
-	"INTEROP_TIMEOUT":    flag.String("interop_timeout", "10", "time limit in seconds on http requests to interop server"),
-	"INTEROP_RETRY_TIME": flag.String("interop_retry_time", "5", "how many seconds to wait after unsuccessful interop authentication"),
-	"INTEROP_MISSION_ID": flag.String("interop_mission_id", "1", "id of the mission assigned to us by the judges"),
-	"RTPP_IP":            flag.String("rtpp_ip", "127.0.0.1", "ip of rtpp computer"),
-	"RTPP_PORT":          flag.String("rtpp_port", "5010", "port of rtpp computer"),
-	"RTPP_TIMEOUT":       flag.String("rtpp_timeout", "360", "time limit in seconds on http requests to interop server"),
-	"RTPP_RETRY_TIME":    flag.String("rtpp_retry_time", "5", "how many seconds to wait after unsuccessful rtpp connection"),
-	"MAV_DEVICE":         flag.String("mav_device", "serial:/dev/serial", "serial port or tcp address of plane to receive messages from"),
-	"MAV_OUTPUT1":        flag.String("mav_output1", "", "first output of mavlink messages"),
-	"MAV_OUTPUT2":        flag.String("mav_output2", "", "second output of mavlink messages"),
-	"MAV_OUTPUT3":        flag.String("mav_output3", "", "third output of mavlink messages"),
-	"MAV_OUTPUT4":        flag.String("mav_output4", "", "fourth output of mavlink messages"),
-	"MAV_OUTPUT5":        flag.String("mav_output5", "", "fifth output of mavlink messages"),
-	"INFLUXDB_URI":       flag.String("influxdb_uri", "http://influxdb:8086", "uri of inlux database for mavlink messages"),
-	"INFLUXDB_TOKEN":     flag.String("influxdb_token", "influxdbToken", "token to allow read/write access to influx database"),
-	"INFLUXDB_BUCKET":    flag.String("influxdb_bucket", "mavlink", "bucket for the influx database"),
-	"INFLUXDB_ORG":       flag.String("influxdb_org", "TritonUAS", "org for the influx database"),
-	"IP":                 flag.String("ip", "*", "ip of interop computer"),
-	"SOCKET_ADDR":        flag.String("socket_addr", "127.0.0.1:6667", "ip + port of path planner zmq"),
-	"DEBUG_MODE":         flag.String("debug", "False", "Boolean to determine logging mode"),
-	"MAV_COMMON_PATH":    flag.String("mav_common_path", "mavlink/message_definitions/v1.0/common.xml", "path to file that contains common mavlink messages"),
-	"MAV_ARDU_PATH":      flag.String("mav_ardu_path", "./mavlink/message_definitions/v1.0/ardupilotmega.xml", "path to file that contains ardupilot mavlink messages"),
+	"HUB_PATH":             flag.String("hub_path", "/home/mat/gopath/src/github.com/tritonuas/hub", "Path to hub folder"),
+	"OBC_ADDR":             flag.String("obc_addr", "127.0.0.1:5010", "ip of obc"),
+	"MAV_DEVICE":           flag.String("mav_device", "serial:/dev/serial", "serial port or tcp address of plane to receive messages from"),
+	"MAV_OUTPUT1":          flag.String("mav_output1", "", "first output of mavlink messages"),
+	"MAV_OUTPUT2":          flag.String("mav_output2", "", "second output of mavlink messages"),
+	"MAV_OUTPUT3":          flag.String("mav_output3", "", "third output of mavlink messages"),
+	"MAV_OUTPUT4":          flag.String("mav_output4", "", "fourth output of mavlink messages"),
+	"MAV_OUTPUT5":          flag.String("mav_output5", "", "fifth output of mavlink messages"),
+	"INFLUXDB_URI":         flag.String("influxdb_uri", "http://influxdb:8086", "uri of inlux database for mavlink messages"),
+	"INFLUXDB_TOKEN":       flag.String("influxdb_token", "influxdbToken", "token to allow read/write access to influx database"),
+	"INFLUXDB_BUCKET":      flag.String("influxdb_bucket", "mavlink", "bucket for the influx database"),
+	"INFLUXDB_ORG":         flag.String("influxdb_org", "TritonUAS", "org for the influx database"),
+	"DEBUG_MODE":           flag.String("debug", "False", "Boolean to determine logging mode"),
+	"ANTENNA_TRACKER_IP":   flag.String("antenna_tracker_ip", "192.168.1.9", "ip address of antenna tracker arduino"),
+	"ANTENNA_TRACKER_PORT": flag.String("antenna_tracker_port", "4000", "port of antenna tracker arduino"),
+	"HOUSTON_PATH":         flag.String("houston_path", "../houston2", "Path to Houston files"),
 }
 
 // setEnvVars will check for any hub related environment variables and
@@ -62,77 +53,51 @@ func setEnvVars() {
 // that all logs go through a central logger.
 // Add in other loggers for modules as needed
 func setLoggers() {
-	ic.Log = log
-	hs.Log = log
 	mav.Log = log
 }
 
-func main() {
+// setupEverything calls all the helper functions to set up the loggers,
+// environment vars, debug mode...
+func setupEverything() {
 	setLoggers()
 	setEnvVars()
 	// prioritize command line flags over environment variables
 	flag.Parse()
-
 	if *ENVS["DEBUG_MODE"] == "True" {
 		log.SetLevel(logrus.DebugLevel)
 		log.Debug("Logging Mode: DEBUG")
 	}
+}
 
-	// create client to interop
-	interopRetryTime, _ := strconv.Atoi(*ENVS["INTEROP_RETRY_TIME"])
-	interopTimeout, _ := strconv.Atoi(*ENVS["INTEROP_TIMEOUT"])
-	interopURL := fmt.Sprintf("%s:%s", *ENVS["INTEROP_IP"], *ENVS["INTEROP_PORT"])
-	interopChannel := make(chan *ic.Client)
-	go ic.EstablishInteropConnection(interopRetryTime, interopURL, *ENVS["INTEROP_USER"], *ENVS["INTEROP_PASS"], interopTimeout, interopChannel)
+func main() {
+	setupEverything()
 
-	// create client to rtpp
-	rtppRetryTime, _ := strconv.Atoi(*ENVS["RTPP_RETRY_TIME"])
-	rtppTimeout, _ := strconv.Atoi(*ENVS["RTPP_TIMEOUT"])
-	rtppURL := fmt.Sprintf("%s:%s", *ENVS["RTPP_IP"], *ENVS["RTPP_PORT"])
-	rtppChannel := make(chan *pp.Client)
-	go pp.EstablishRTPPConnection(rtppRetryTime, rtppURL, rtppTimeout, rtppChannel)
-
-	// Do other things...
-	telemetryChannel := make(chan *ic.Telemetry, 100)
-	sendWaypointToPlaneChannel := make(chan *pp.Path)
-
-	// begins to send messages from the plane to InfluxDB
-	mavOutputs := []string{*ENVS["MAV_OUTPUT1"], *ENVS["MAV_OUTPUT2"], *ENVS["MAV_OUTPUT3"], *ENVS["MAV_OUTPUT4"], *ENVS["MAV_OUTPUT5"]}
-	go mav.RunMavlink(
-		*ENVS["MAV_COMMON_PATH"],
-		*ENVS["MAV_ARDU_PATH"],
-		*ENVS["INFLUXDB_TOKEN"],
-		*ENVS["INFLUXDB_BUCKET"],
-		*ENVS["INFLUXDB_ORG"],
-		*ENVS["MAV_DEVICE"],
-		*ENVS["INFLUXDB_URI"],
-		mavOutputs,
-		telemetryChannel,
-		sendWaypointToPlaneChannel)
-
-	var server *hs.Server
-	server = new(hs.Server)
-
-	port := "5000"
-	log.Infof("Creating Hub server at port %s", port)
-	interopMissionID, err := strconv.Atoi(*ENVS["INTEROP_MISSION_ID"])
-	if err != nil {
-		log.Warningf("Invalid interop mission id (cannot parse \"%s\" to int). Using default mission id 1.", *ENVS["INTEROP_MISSION_ID"])
-		interopMissionID = 1
-	} else {
-		log.Infof("Using interop mission id of %d", interopMissionID)
+	influxCreds := influxdb.Credentials{
+		Token:  *ENVS["INFLUXDB_TOKEN"],
+		Bucket: *ENVS["INFLUXDB_BUCKET"],
+		Org:    *ENVS["INFLUXDB_ORG"],
+		URI:    *ENVS["INFLUXDB_URI"],
 	}
 
-	server.Run(
-		port,
-		interopChannel,
-		interopMissionID,
-		rtppChannel,
-		telemetryChannel,
-		*ENVS["INFLUXDB_URI"],
-		*ENVS["INFLUXDB_TOKEN"],
-		*ENVS["INFLUXDB_BUCKET"],
-		*ENVS["INFLUXDB_ORG"],
-		sendWaypointToPlaneChannel)
+	influxClient := influxdb.New(influxCreds)
 
+	mavlinkClient := mav.New(
+		influxClient,
+		*ENVS["ANTENNA_TRACKER_IP"],
+		*ENVS["ANTENNA_TRACKER_PORT"],
+		*ENVS["MAV_DEVICE"],
+		*ENVS["MAV_OUTPUT1"],
+		*ENVS["MAV_OUTPUT2"],
+		*ENVS["MAV_OUTPUT3"],
+		*ENVS["MAV_OUTPUT4"],
+		*ENVS["MAV_OUTPUT5"],
+	)
+
+	obcClient := obc.NewClient(*ENVS["OBC_ADDR"], 60)
+
+	go mavlinkClient.Listen()
+
+	// Set up GIN HTTP Server
+	server := server.New(influxClient, mavlinkClient, obcClient)
+	server.Start()
 }
